Add delete method
This method implements the logic for deleting an element at a user-specified position. There is also a check to make sure that the specified index is valid, i.e. not greater than the length of the list and not a negative value. If the element is deleted correctly, the array size is reduced.