Add deleteAll method
When calling this method, you must specify the value of the element that you want to remove from the list. Not one, but all such elements will be removed. If you enter an incorrect value for an element, the contents of the list will not change. If an element is removed correctly, the size of the list is reduced by one.