Add base for CircularLinkedList and connect to tests
This commit adds base for further implementation of functionality for a singly linked circular list. A file with the implementation of the logic for this list has also been attached to the tests. Certain tests had to be committed due to the lack of implementation of the logic of certain methods at the moment. As the logic is implemented, the tests will be decommitted.